Do img mask float conversion at train and test

+++ scripts/data_setup.py
@@ -72,9 +72,9 @@ class LettuceSegDataset(Dataset):
         # Retrieve image and mask, should be np.array for albumentations.transforms
         img = np.array(Image.open(self.img_paths[index]))
         if self.mask_paths[index].endswith(".json"):
-            mask = utils.binary_poly2px(self.mask_paths[index]).astype(np.float32)
+            mask = utils.binary_poly2px(self.mask_paths[index])
         else:
-            mask = np.array(Image.open(self.mask_paths[index]), dtype=np.float32)
+            mask = np.array(Image.open(self.mask_paths[index]))
         if 255.0 in mask:
             mask[mask == 255.0] = 1.0
